+++ test/tls_certificate_check_dependent_app_SUITE_data/dependent_app/apps/dependent_app/src/dependent_app.erl
@@ -1,0 +1,3 @@
+-module(dependent_app).
+
+-export([]).