+++ CHANGELOG.md
@@ -18,6 +18,12 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - module with bundled CAs to latest as of 2025/11/04, 04:12 UTC
 (source: https://curl.se/ca/cacert.pem)
 
+### Fixed
+
+- throwing of `application_either_not_started_or_not_ready' when this lib is
+  used as an _implicit_ dependency (i.e. called before it has started or
+  finished initializing)
+
 ## [1.29.0] - 2025-08-14
 
 ### Added